+++ akka-remote/src/main/resources/reference.conf
@@ -38,29 +38,82 @@ akka {
 
   remote {
 
-    # Which implementation of akka.remote.RemoteSupport to use
+    # Which implementation of akka.remote.RemoteTransport to use
     # default is a TCP-based remote transport based on Netty
-    transport = "akka.remote.netty.NettyRemoteSupport"
+    transport = "akka.remote.netty.NettyRemoteTransport"
 
-    # In case of increased latency / overflow how long
-    # should we wait (blocking the sender) until we deem the send to be cancelled?
-    # 0 means "never backoff", any positive number will indicate time to block at most.
-    backoff-timeout = 0ms
-
-    use-compression = off
-
-    # Generate your own with '$AKKA_HOME/scripts/generate_config_with_secure_cookie.sh'
-    #     or using 'akka.util.Crypt.generateSecureCookie'
-    secure-cookie = ""
+    # Enable untrusted mode for full security of server managed actors, allows
+    # untrusted clients to connect.
+    untrusted-mode = off
 
     # Timeout for ACK of cluster operations, lik checking actor out etc.
     remote-daemon-ack-timeout = 30s
 
-    # Reuse inbound connections for outbound messages
-    use-passive-connections = on
+    # Each property is annotated with (C) or (S) or (C&S), where C stands for “client” and S for “server” role.
+    # The NettyRemoteTransport always starts the server role to allow inbound connections, and it starts
+    # active client connections whenever sending to a destination which is not yet connected; if configured
+    # it reuses inbound connections for replies, which is called a passive client connection (i.e. from server
+    # to client).
+    netty {
+      
+      # (C) In case of increased latency / overflow how long
+      # should we wait (blocking the sender) until we deem the send to be cancelled?
+      # 0 means "never backoff", any positive number will indicate time to block at most.
+      backoff-timeout = 0ms
+  
+      # (C&S) Generate your own with '$AKKA_HOME/scripts/generate_config_with_secure_cookie.sh'
+      #     or using 'akka.util.Crypt.generateSecureCookie'
+      secure-cookie = ""
+  
+      # (S) Should the remote server require that it peers share the same secure-cookie
+      # (defined in the 'remote' section)?
+      require-cookie = off
 
-    # Whether any Threds created by the remoting should be daemons or not
-    daemonic = on
+      # (S) Reuse inbound connections for outbound messages
+      use-passive-connections = on
+  
+      # (C&S) Whether any Threds created by the remoting should be daemons or not
+      daemonic = on
+
+      # (S) The hostname or ip to bind the remoting to,
+      # InetAddress.getLocalHost.getHostAddress is used if empty
+      hostname = ""
+
+      # (S) The default remote server port clients should connect to.
+      # Default is 2552 (AKKA), use 0 if you want a random available port
+      port = 2552
+
+      # (C&S) Increase this if you want to be able to send messages with large payloads
+      message-frame-size = 1 MiB
+
+      # (C) Timeout duration
+      connection-timeout = 120s
+
+      # (S) Sets the size of the connection backlog
+      backlog = 4096
+
+      # (S) Length in akka.time-unit how long core threads will be kept alive if idling
+      execution-pool-keepalive = 60s
+
+      # (S) Size of the core pool of the remote execution unit
+      execution-pool-size = 4
+
+      # (S) Maximum channel size, 0 for off
+      max-channel-memory-size = 0b
+
+      # (S) Maximum total size of all channels, 0 for off
+      max-total-memory-size = 0b
+
+      # (C) Time between reconnect attempts for active clients
+      reconnect-delay = 5s
+
+      # (C) Inactivity period after which active client connection is shutdown; will be
+      # re-established in case of new communication requests
+      read-timeout = 3600s
+
+      # (C) Maximum time window that a client should try to reconnect for
+      reconnection-time-window = 600s
+    }
 
     # accrual failure detection config
     failure-detector {
@@ -93,52 +146,6 @@ akka {
       daemonic = on
     }
 
-    server {
-      # The hostname or ip to bind the remoting to,
-      # InetAddress.getLocalHost.getHostAddress is used if empty
-      hostname = ""
-
-      # The default remote server port clients should connect to.
-      # Default is 2552 (AKKA), use 0 if you want a random available port
-      port = 2552
-
-      # Increase this if you want to be able to send messages with large payloads
-      message-frame-size = 1 MiB
-
-      # Timeout duration
-      connection-timeout = 120s
-
-      # Should the remote server require that it peers share the same secure-cookie
-      # (defined in the 'remote' section)?
-      require-cookie = off
-
-      # Enable untrusted mode for full security of server managed actors, allows
-      # untrusted clients to connect.
-      untrusted-mode = off
-
-      # Sets the size of the connection backlog
-      backlog = 4096
-
-      # Length in akka.time-unit how long core threads will be kept alive if idling
-      execution-pool-keepalive = 60s
-
-      # Size of the core pool of the remote execution unit
-      execution-pool-size = 4
-
-      # Maximum channel size, 0 for off
-      max-channel-memory-size = 0b
-
-      # Maximum total size of all channels, 0 for off
-      max-total-memory-size = 0b
-    }
-
-    client {
-      reconnect-delay = 5s
-      read-timeout = 3600s
-      message-frame-size = 1 MiB
-      # Maximum time window that a client should try to reconnect for
-      reconnection-time-window = 600s
-    }
   }
 
   cluster {
