+++ akka-remote/src/main/scala/akka/remote/RemoteAddress.scala
@@ -1,0 +1,5 @@
+/**
+ *  Copyright (C) 2009-2012 Typesafe Inc. <http://www.typesafe.com>
+ */
+package akka.remote
+
