+++ akka-remote/src/multi-jvm/scala/akka/remote/AkkaRemoteSpec.scala
@@ -32,14 +32,4 @@ abstract class AkkaRemoteSpec(config: Config)
   extends AkkaSpec(config.withFallback(AkkaRemoteSpec.testConf))
   with MultiJvmSync {
 
-  /**
-   * Helper function for accessing the underlying remoting.
-   */
-  def remote: Remote = {
-    system.asInstanceOf[ActorSystemImpl].provider match {
-      case r: RemoteActorRefProvider ⇒ r.remote
-      case _                         ⇒ throw new Exception("Remoting is not enabled")
-    }
-  }
-
 }
