+++ akka-remote/src/test/scala/akka/remote/RemoteCommunicationSpec.scala
@@ -33,7 +33,7 @@ object RemoteCommunicationSpec {
 class RemoteCommunicationSpec extends AkkaSpec("""
 akka {
   actor.provider = "akka.remote.RemoteActorRefProvider"
-  remote.server {
+  remote.netty {
     hostname = localhost
     port = 12345
   }
@@ -47,7 +47,7 @@ akka {
 
   import RemoteCommunicationSpec._
 
-  val conf = ConfigFactory.parseString("akka.remote.server.port=12346").withFallback(system.settings.config)
+  val conf = ConfigFactory.parseString("akka.remote.netty.port=12346").withFallback(system.settings.config)
   val other = ActorSystem("remote_sys", conf)
 
   val remote = other.actorOf(Props(new Actor {
