+++ akka-remote/src/test/scala/akka/remote/RemoteRouterSpec.scala
@@ -20,7 +20,7 @@ object RemoteRouterSpec {
 class RemoteRouterSpec extends AkkaSpec("""
 akka {
   actor.provider = "akka.remote.RemoteActorRefProvider"
-  remote.server {
+  remote.netty {
     hostname = localhost
     port = 12345
   }
@@ -44,7 +44,7 @@ akka {
 
   import RemoteRouterSpec._
 
-  val conf = ConfigFactory.parseString("akka.remote.server.port=12346").withFallback(system.settings.config)
+  val conf = ConfigFactory.parseString("akka.remote.netty.port=12346").withFallback(system.settings.config)
   val other = ActorSystem("remote_sys", conf)
 
   override def atTermination() {
